+++ src/sync_ftp/data_stream.rs
@@ -14,7 +14,7 @@ use std::net::TcpStream;
 pub enum DataStream {
     Tcp(TcpStream),
     #[cfg(feature = "secure")]
-    Ssl(TlsStream<TcpStream>),
+    Ssl(TlsStreamWrapper),
 }
 
 #[cfg(feature = "secure")]
@@ -25,7 +25,7 @@ impl DataStream {
     pub fn into_tcp_stream(self) -> TcpStream {
         match self {
             DataStream::Tcp(stream) => stream,
-            DataStream::Ssl(stream) => stream.get_ref().try_clone().unwrap(),
+            DataStream::Ssl(stream) => stream.tcp_stream(),
         }
     }
 
@@ -57,7 +57,7 @@ impl Read for DataStream {
         match self {
             DataStream::Tcp(ref mut stream) => stream.read(buf),
             #[cfg(feature = "secure")]
-            DataStream::Ssl(ref mut stream) => stream.read(buf),
+            DataStream::Ssl(ref mut stream) => stream.mut_ref().read(buf),
         }
     }
 }
@@ -67,7 +67,7 @@ impl Write for DataStream {
         match self {
             DataStream::Tcp(ref mut stream) => stream.write(buf),
             #[cfg(feature = "secure")]
-            DataStream::Ssl(ref mut stream) => stream.write(buf),
+            DataStream::Ssl(ref mut stream) => stream.mut_ref().write(buf),
         }
     }
 
@@ -75,7 +75,67 @@ impl Write for DataStream {
         match self {
             DataStream::Tcp(ref mut stream) => stream.flush(),
             #[cfg(feature = "secure")]
-            DataStream::Ssl(ref mut stream) => stream.flush(),
+            DataStream::Ssl(ref mut stream) => stream.mut_ref().flush(),
+        }
+    }
+}
+
+// -- tls stream wrapper to implement drop...
+
+#[cfg(feature = "secure")]
+#[derive(Debug)]
+/// Tls stream wrapper. This type is a garbage data type used to impl the drop trait for the tls stream.
+/// This allows me to keep returning `Read` and `Write` traits in stream methods
+pub struct TlsStreamWrapper {
+    stream: TlsStream<TcpStream>,
+    ssl_shutdown: bool,
+}
+
+#[cfg(feature = "secure")]
+impl TlsStreamWrapper {
+    /// Get underlying tcp stream
+    pub(crate) fn tcp_stream(mut self) -> TcpStream {
+        let mut stream = self.stream.get_ref().try_clone().unwrap();
+        // Don't perform shutdown later
+        self.ssl_shutdown = false;
+        // flush stream (otherwise can cause bad chars on channel)
+        if let Err(err) = stream.flush() {
+            error!("Error in flushing tcp stream: {}", err);
+        }
+        trace!("TLS stream terminated");
+        stream
+    }
+
+    /// Get ref to underlying tcp stream
+    pub(crate) fn get_ref(&self) -> &TcpStream {
+        self.stream.get_ref()
+    }
+
+    /// Get mutable reference to tls stream
+    pub(crate) fn mut_ref(&mut self) -> &mut TlsStream<TcpStream> {
+        &mut self.stream
+    }
+}
+
+#[cfg(feature = "secure")]
+impl From<TlsStream<TcpStream>> for TlsStreamWrapper {
+    fn from(stream: TlsStream<TcpStream>) -> Self {
+        Self {
+            stream,
+            ssl_shutdown: true,
+        }
+    }
+}
+
+#[cfg(feature = "secure")]
+impl Drop for TlsStreamWrapper {
+    fn drop(&mut self) {
+        if self.ssl_shutdown {
+            if let Err(err) = self.stream.shutdown() {
+                error!("Failed to shutdown stream: {}", err);
+            } else {
+                debug!("TLS Stream shut down");
+            }
         }
     }
 }